Move the themes tab
