+++ src/lib/components/HubVersionCheck.svelte
@@ -54,7 +54,7 @@
 </script>
 
 
-<section class="card w-full">
+<section class="card hubupdatemessage">
     <div class="p-4 space-y-4">
         <div class="flex gap-4">
             {#if latestVersionErrorFound === true}
@@ -63,7 +63,6 @@
                 <button type="button" on:click={TryFetchingLatestVersion} class="text-white bg-red-700 hover:bg-red-800 font-medium rounded-lg text-sm px-3 py-1 dark:bg-red-600 dark:hover:bg-red-700 button-red"><i class="fa-solid fa-triangle-exclamation"></i> Retry</button>
             {:else if latestVersion === "Loading..."}
                 <span>Checking for updates...</span>
-                <div class="placeholder animate-pulse flex-1" />
             {:else if VersionHub != latestVersion}
                 <span><b>Your OpenTaiko Hub installation is out of date.</b></span>
                 <button type="button" on:click={UpdateHub} class="text-white bg-green-700 hover:bg-green-800 font-medium rounded-lg text-sm px-3 py-1 dark:bg-green-600 dark:hover:bg-green-700 button-green"><i class="fa-solid fa-download"></i> Update OpenTaiko Hub to {latestVersion.slice(1)}</button>
@@ -80,7 +79,12 @@
     .hubversionnumber {
         font-weight: bold;
         position: fixed;
-        bottom: 10px;
-        right: 15px;
+        bottom: 16px;
+        left: 88px;
+    }
+    .hubupdatemessage {
+        position: fixed;
+        bottom: 16px;
+        right: 16px;
     }
 </style>
